Add some CruiseControl.NET enumeration classes
Add some enumeration classes which are partly copied from the official CC.Net source code on GitHub (https://github.com/ccnet/CruiseControl.NET) and updated/renamed by hand. This prevents the reference of any of the official NuGet packages or dll files, since that would cause an overhead due to the small functionality of this widget compared to the CCTray application and similar tools.