Add some CruiseControl.NET model classes
Add some model classes to map the data from the XML data retrieved to plain C# classes. Parts of these classes are copied from the official CC.Net source code on GitHub (https://github.com/ccnet/CruiseControl.NET) and updated/renamed by hand. This prevents the reference of any of the official NuGet packages or dll files, since that would cause an overhead due to the small functionality of this widget compared to the CCTray application and similar tools.