Create basic build widget and health check
Create basic configurable widget to get project name and server URL from the user which is then used to request the state of the project from the CruiseControl.NET build server. The actual health check is implemented using a simple .NET HTTP client, which queries the XML status report from the server and updates the health state of the widget accordingly.