Simplify health check implementation
Make use of the new extension method to convert the integration state retrieved from the CC.NET server to the AnyStatus State, so that the entire switch-statement is replaced. Use the new instance of the XmlSerializer which has been wired up to provide more information for Debug builds instead of creating a second brand new instance to actually parse the server's response.