Remove URL validation and update internal properties
Remove the over-engineered validation of the supplied URL because the input field is cleared way too often which does not help at all due to the fact that no indication is shown why this has happened. Instead, allow any kind of valid URL and let the user click the 'Test' button on the UI if validation of arguments is desired. The setter of property
'URL' does now set the values for properties 'ViewFarmReportUrl' and 'XmlStatusReportUrl' as well by stripping the query (if set) and the last segment of the supplied URL to build a valid URL.