Improve health check
Improve implementation of IHealthCheck so that no more unnecessary notifications will be shown because property State will only be set once. This prevents overwriting property PreviousState, which in turn lets the default implementation of IsNotificationRequired() handle this accordingly.
Additionally, multiple new data will be stored within the widget. For example, the dashboard URL of the project is contained within the XML status and will be used later to open this exact web page in the user's web browser. Additionally, the ServerName used to force start an integration build is now stored, too.